+++ Kurkku/Messages/Outgoing/Catalogue/CatalogItemDiscountComposer.cs
@@ -14,8 +14,8 @@ namespace Kurkku.Messages.Outgoing
         public override void Write()
         {
             m_Data.Add(discount.PurchaseLimit); // The discount / bulk buy limit
-            m_Data.Add(discount.ItemCountRequired); // A - "Buy A get B free"
-            m_Data.Add(discount.ItemCountFree); // B
+            m_Data.Add((int)discount.ItemCountRequired); // A - "Buy A get B free"
+            m_Data.Add((int)discount.ItemCountFree); // B
             m_Data.Add(0);
             m_Data.Add(0);//Count
             /*{
